fabs function requires a double argument

+++ magick/cache.c
@@ -3590,7 +3590,7 @@ static inline void ApplyPixelCompositeMask(const MagickPixelPacket *p,
   double
     gamma;
 
-  if (fabs(alpha-TransparentOpacity) < MagickEpsilon)
+  if (fabs((double) (alpha-TransparentOpacity)) < MagickEpsilon)
     {
       *composite=(*q);
       return;
